+++ test/phoenix_test/playwright/multiple_browsers_parameterize_test.exs
@@ -3,9 +3,11 @@ defmodule PhoenixTest.Playwright.MultipleBrowsersParameterizeTest do
     async: true,
     parameterize: [%{browser_pool: :chromium_pool}, %{browser_pool: :firefox_pool}]
 
+  import PhoenixTest.TestHelpers, only: [set_html: 2]
+
   test "run the same test in multiple browsers (checkout from pools)", %{conn: conn} do
     conn
-    |> visit("/pw/live")
+    |> set_html("<h1>Parameterized browser</h1>")
     |> assert_has("h1")
   end
 end