+++ test/support/playwright/page_controller.ex
@@ -1,20 +1,10 @@
 defmodule PhoenixTest.Playwright.PageController do
   use Phoenix.Controller, formats: [html: "View"]
 
-  def other(conn, _) do
-    render(conn, "other.html")
-  end
-
-  def longer_than_viewport(conn, _) do
-    render(conn, "longer_than_viewport.html")
-  end
+  def blank(conn, _), do: send_resp(conn, 200, "")
 
   def download(conn, _) do
     send_download(conn, {:file, "test/files/elixir.jpg"})
-  end
-
-  def download_link(conn, _) do
-    render(conn, "download_link.html")
   end
 
   def js_script_console_error(conn, _) do