+++ test/support/playwright/page_view.ex
@@ -1,26 +1,6 @@
 defmodule PhoenixTest.Playwright.PageView do
   use Phoenix.Component
 
-  def render("other.html", assigns) do
-    ~H"""
-    <h1>Other</h1>
-    """
-  end
-
-  def render("longer_than_viewport.html", assigns) do
-    ~H"""
-    <h1>Longer than viewport</h1>
-    <div :for={_ <- 1..100}>Lorem ipsum</div>
-    """
-  end
-
-  def render("download_link.html", assigns) do
-    ~H"""
-    <h1>Download link</h1>
-    <a href="/pw/download">Download image</a>
-    """
-  end
-
   def render("js_script_console_error.html", assigns) do
     ~H"""
     <script type="text/javascript">