+++ test/support/upstream/web_app/router.ex
@@ -32,12 +32,9 @@ defmodule PhoenixTest.WebApp.Router do
   scope "/pw", PhoenixTest.Playwright do
     pipe_through([:browser, :csrf])
 
-    live("/live", Live)
     live("/live/ecto", EctoLive)
+    get("/blank", PageController, :blank)
     get("/download", PageController, :download)
-    get("/download-link", PageController, :download_link)
-    get("/other", PageController, :other)
-    get("/longer-than-viewport", PageController, :longer_than_viewport)
     get("/cookies", PageController, :cookies)
     get("/session", PageController, :session)
     get("/headers", PageController, :headers)